kleine Anpassungen. Zb Port 16output um uP-speed test.

+++ Grid.py
@@ -21,8 +21,8 @@ class Grid():
         unpaintedObstacles=obstacles
         
         for obstacle in unpaintedObstacles:
-            obstacle_x_grid=int(obstacle[0]/10)
-            obstacle_y_grid=int(obstacle[1]/10)
+            obstacle_x_grid=(obstacle[0]/10)
+            obstacle_y_grid=(obstacle[1]/10)
             
             if (obstacle_x_grid,obstacle_y_grid) not in  self.walls:
                 self.walls.append((obstacle_x_grid,obstacle_y_grid))
